requirements.txt
rich>=13.0.0
pyusb>=1.2.1
crccheck>=1.3.0 
setuptools